+++ lte/cloud/go/services/subscriberdb/const.go
@@ -19,7 +19,7 @@ const (
 	EntityType = "subscriber"
 
 	LookupTableBlobstore     = "subscriber_lookup_blobstore"
-	SyncstoreBlobstore       = "subscriber_syncstore_blobstore"
+	SyncstoreTableBlobstore  = "subscriber_syncstore_blobstore"
 	SyncstoreTableNamePrefix = "subscriber"
 
 	// MinimumSyncInterval is the the minimum interval in seconds between
