+++ lte/cloud/go/services/subscriberdb/digest_test.go
@@ -374,86 +374,3 @@ func TestGetPerSubscriberDigests(t *testing.T) {
 	assert.Equal(t, "IMSI00003", perSubDigests[1].Id)
 	assert.NotEqual(t, digestSub3, perSubDigests[1].Digest.Md5Base64Digest)
 }
-
-func TestGetLeafDigestsDiff(t *testing.T) {
-	t.Run("both empty", func(t *testing.T) {
-		prev, next := []*protos.LeafDigest{}, []*protos.LeafDigest{}
-		toRenew, deleted := subscriberdb.GetLeafDigestsDiff(prev, next)
-		assert.Empty(t, toRenew)
-		assert.Empty(t, deleted)
-	})
-	t.Run("all empty", func(t *testing.T) {
-		prev := []*protos.LeafDigest{
-			{Id: "IMSI00000", Digest: &protos.Digest{Md5Base64Digest: "apple"}},
-			{Id: "IMSI00001", Digest: &protos.Digest{Md5Base64Digest: "banana"}},
-			{Id: "IMSI00002", Digest: &protos.Digest{Md5Base64Digest: "cherry"}},
-		}
-
-		next := []*protos.LeafDigest{}
-		toRenew, deleted := subscriberdb.GetLeafDigestsDiff(prev, next)
-
-		assert.Empty(t, toRenew)
-		assert.Equal(t, []string{"IMSI00000", "IMSI00001", "IMSI00002"}, deleted)
-	})
-	t.Run("tracked empty", func(t *testing.T) {
-		prev := []*protos.LeafDigest{}
-		next := []*protos.LeafDigest{
-			{Id: "IMSI00000", Digest: &protos.Digest{Md5Base64Digest: "apple"}},
-			{Id: "IMSI00001", Digest: &protos.Digest{Md5Base64Digest: "banana"}},
-			{Id: "IMSI00002", Digest: &protos.Digest{Md5Base64Digest: "cherry"}},
-		}
-		expectedToRenew := map[string]string{
-			"IMSI00000": "apple",
-			"IMSI00001": "banana",
-			"IMSI00002": "cherry",
-		}
-		toRenew, deleted := subscriberdb.GetLeafDigestsDiff(prev, next)
-		assert.Equal(t, expectedToRenew, toRenew)
-		assert.Empty(t, deleted)
-	})
-	t.Run("both not empty, basic", func(t *testing.T) {
-		prev := []*protos.LeafDigest{
-			{Id: "IMSI00001", Digest: &protos.Digest{Md5Base64Digest: "apple"}},
-			{Id: "IMSI00002", Digest: &protos.Digest{Md5Base64Digest: "banana"}},
-			{Id: "IMSI00004", Digest: &protos.Digest{Md5Base64Digest: "dragonfruit"}},
-		}
-		next := []*protos.LeafDigest{
-			{Id: "IMSI00001", Digest: &protos.Digest{Md5Base64Digest: "apple"}},
-			{Id: "IMSI00002", Digest: &protos.Digest{Md5Base64Digest: "banana2"}},
-			{Id: "IMSI00003", Digest: &protos.Digest{Md5Base64Digest: "cherry"}},
-		}
-		expectedToRenew := map[string]string{
-			"IMSI00002": "banana2",
-			"IMSI00003": "cherry",
-		}
-		toRenew, deleted := subscriberdb.GetLeafDigestsDiff(prev, next)
-		assert.Equal(t, expectedToRenew, toRenew)
-		assert.Equal(t, []string{"IMSI00004"}, deleted)
-	})
-	t.Run("both not empty, involved", func(t *testing.T) {
-		prev := []*protos.LeafDigest{
-			{Id: "IMSI00001", Digest: &protos.Digest{Md5Base64Digest: "apple"}},
-			{Id: "IMSI00002", Digest: &protos.Digest{Md5Base64Digest: "banana"}},
-			{Id: "IMSI00004", Digest: &protos.Digest{Md5Base64Digest: "dragonfruit"}},
-			{Id: "IMSI00006", Digest: &protos.Digest{Md5Base64Digest: "fig"}},
-			{Id: "IMSI00007", Digest: &protos.Digest{Md5Base64Digest: "grape"}},
-		}
-		next := []*protos.LeafDigest{
-			{Id: "IMSI00001", Digest: &protos.Digest{Md5Base64Digest: "apple"}},
-			{Id: "IMSI00002", Digest: &protos.Digest{Md5Base64Digest: "banana"}},
-			{Id: "IMSI00003", Digest: &protos.Digest{Md5Base64Digest: "cherry"}},
-			{Id: "IMSI00005", Digest: &protos.Digest{Md5Base64Digest: "eggplant"}},
-			{Id: "IMSI00006", Digest: &protos.Digest{Md5Base64Digest: "fig2"}},
-			{Id: "IMSI00008", Digest: &protos.Digest{Md5Base64Digest: "honeydew"}},
-		}
-		expectedToRenew := map[string]string{
-			"IMSI00003": "cherry",
-			"IMSI00005": "eggplant",
-			"IMSI00006": "fig2",
-			"IMSI00008": "honeydew",
-		}
-		toRenew, deleted := subscriberdb.GetLeafDigestsDiff(prev, next)
-		assert.Equal(t, expectedToRenew, toRenew)
-		assert.Equal(t, []string{"IMSI00004", "IMSI00007"}, deleted)
-	})
-}
