+++ lte/cloud/go/services/subscriberdb_cache/worker_test.go
@@ -144,7 +144,7 @@ func TestSubscriberdbCacheWorker(t *testing.T) {
 	subProtos, nextToken, err = store.GetCachedByPage("n1", "", 2)
 	assert.NoError(t, err)
 	assert.Len(t, subProtos, 2)
-	subProtosDeserialized, err := subscriberdb.DeserializeSubProtos(subProtos)
+	subProtosDeserialized, err := subscriberdb.DeserializeSubscribers(subProtos)
 	assert.NoError(t, err)
 	assert.True(t, proto.Equal(sub1, subProtosDeserialized[0]))
 	assert.True(t, proto.Equal(sub2, subProtosDeserialized[1]))
@@ -216,7 +216,7 @@ func TestUpdateSubProtosByNetworkNoChange(t *testing.T) {
 	page, _, err := store.GetCachedByPage("n1", "", 3)
 	assert.NoError(t, err)
 	assert.Len(t, page, 3)
-	subProtos, err := subscriberdb.DeserializeSubProtos(page)
+	subProtos, err := subscriberdb.DeserializeSubscribers(page)
 	assert.NoError(t, err)
 	assert.True(t, proto.Equal(subProtoFromID("IMSI00001"), subProtos[0]))
 	assert.True(t, proto.Equal(subProtoFromID("IMSI00002"), subProtos[1]))
@@ -265,7 +265,7 @@ func subProtoFromID(sid string) *lte_protos.SubscriberData {
 func initializeSyncstore(t *testing.T) syncstore.SyncStore {
 	db, err := test_utils.GetSharedMemoryDB()
 	assert.NoError(t, err)
-	fact := blobstore.NewSQLBlobStorageFactory(subscriberdb.SyncstoreBlobstore, db, sqorc.GetSqlBuilder())
+	fact := blobstore.NewSQLBlobStorageFactory(subscriberdb.SyncstoreTableBlobstore, db, sqorc.GetSqlBuilder())
 	assert.NoError(t, fact.InitializeFactory())
 	store, err := syncstore.NewSyncStore(db, sqorc.GetSqlBuilder(), fact, syncstore.Config{TableNamePrefix: "subscriber", CacheWriterValidIntervalSecs: 150})
 	assert.NoError(t, err)
