+++ orc8r/cloud/go/orc8r/math/math.go
@@ -15,30 +15,34 @@ package math
 
 import "hash/fnv"
 
-// JitterUint32 returns a jitter of the given uint32 value that is deterministic
-// based on the given key.
-func JitterUint32(n uint32, key string, maxMultiplier float32) uint32 {
+func getFNV1Hash(key string) uint32 {
 	// FNV-1 is a non-cryptographic hash function that is fast and very simple to implement
 	h := fnv.New32a()
 	_, err := h.Write([]byte(key))
 	if err != nil {
 		return 0
 	}
-	multiplier := float32(h.Sum32()%100) / 100.0
-	maxJitter := float32(n) * maxMultiplier
-	return uint32(multiplier * maxJitter)
+	return h.Sum32()
+}
+
+// JitterUint32 returns a jitter of the given uint32 value that is deterministic
+// based on the given key.
+func JitterUint32(n uint32, key string, maxMultiplier float32) uint32 {
+	fnv1Hash := getFNV1Hash(key)
+	multiplier := maxMultiplier * float32(fnv1Hash%100) / 100.0
+	if multiplier >= 1 {
+		return n
+	}
+	return uint32(multiplier * float32(n))
 }
 
 // JitterInt64 returns a jitter of the given int64 value that is deterministic
 // based on the given key.
 func JitterInt64(n int64, key string, maxMultiplier float64) int64 {
-	// FNV-1 is a non-cryptographic hash function that is fast and very simple to implement
-	h := fnv.New32a()
-	_, err := h.Write([]byte(key))
-	if err != nil {
-		return 0
+	fnv1Hash := getFNV1Hash(key)
+	multiplier := maxMultiplier * float64(fnv1Hash%100) / 100.0
+	if multiplier >= 1 {
+		return n
 	}
-	multiplier := float64(h.Sum32()%100) / 100.0
-	maxJitter := float64(n) * maxMultiplier
-	return int64(multiplier * maxJitter)
+	return int64(multiplier * float64(n))
 }
